getDaily emotes and getPlusEmotes are now in one method.
