+++ labynode.js
@@ -1,7 +1,6 @@
 const status = require('./util/getStatus');
 const emotes = require('./util/getEmotes');
-const dailyEmotes = require('./util/getDailyEmotes');
-const plusEmotes = require('./util/getPlusEmotes');
+const shopEmotes = require('./util/getShopEmotes');
 const cosmetics = require('./util/getCosmetics');
 const roles = require('./util/getRoles');
 const stickers = require('./util/getStickers');
@@ -27,11 +26,7 @@ module.exports = {
         await stickers.getStickers(uuid, callback)
     },
 
-    getDailyEmotes: async function getDailyEmotes(callback) {
-        await dailyEmotes.getDailyEmotes(callback);
-    },
-
-    getPlusEmotes: async function getPlusEmotes(callback) {
-        await plusEmotes.getPlusEmotes(callback);
+    getShopEmotes: async function getShopEmotes(type, callback) {
+        await shopEmotes.getShopEmotes(type, callback);
     }
 }
